Place context menu in the layer list
