+++ packages/story-editor/src/components/canvas/framesLayer.js
@@ -27,9 +27,14 @@ import { STORY_ANIMATION_STATE } from '@web-stories-wp/animation';
  * Internal dependencies
  */
 import { DESIGN_SPACE_MARGIN } from '../../constants';
-import { useStory, useCanvas, useLayout, useTransform } from '../../app';
+import {
+  useStory,
+  useCanvas,
+  useLayout,
+  useTransform,
+  useRightClickMenu,
+} from '../../app';
 import useCanvasKeys from '../../app/canvas/useCanvasKeys';
-import { useRightClickMenu } from '../../app/rightClickMenu';
 import PageMenu from './pagemenu';
 import { Layer, MenuArea, NavNextArea, NavPrevArea, PageArea } from './layout';
 import FrameElement from './frameElement';
