+++ packages/story-editor/src/components/panels/design/layer/layerList.js
@@ -18,7 +18,12 @@
  * External dependencies
  */
 import styled from 'styled-components';
-import { Fragment, useCallback } from '@web-stories-wp/react';
+import {
+  Fragment,
+  useCallback,
+  useEffect,
+  useRef,
+} from '@web-stories-wp/react';
 import PropTypes from 'prop-types';
 import { __ } from '@web-stories-wp/i18n';
 
@@ -30,7 +35,7 @@ import {
   ReorderableSeparator,
   ReorderableItem,
 } from '../../../reorderable';
-import { useStory } from '../../../../app';
+import { useRightClickMenu, useStory } from '../../../../app';
 import useFocusCanvas from '../../../canvas/useFocusCanvas';
 import { LAYER_HEIGHT } from './constants';
 import Layer from './layer';
@@ -59,6 +64,9 @@ function LayerPanel({ layers }) {
     setSelectedElementsById: state.actions.setSelectedElementsById,
   }));
 
+  const rightClickAreaRef = useRef(null);
+  const { onOpenMenu } = useRightClickMenu();
+
   const numLayers = layers && layers.length;
 
   const focusCanvas = useFocusCanvas();
@@ -70,12 +78,29 @@ function LayerPanel({ layers }) {
     [setSelectedElementsById, focusCanvas]
   );
 
+  // Override the browser's context menu within this node
+  useEffect(() => {
+    const node = rightClickAreaRef.current;
+    if (!node) {
+      return undefined;
+    }
+
+    node.addEventListener('contextmenu', onOpenMenu);
+
+    return () => {
+      node.removeEventListener('contextmenu', onOpenMenu);
+    };
+    // On initial load there are no layers. Ref update won't trigger effect.
+    // So re-run when layers are fully updated.
+  }, [numLayers, onOpenMenu]);
+
   if (!numLayers) {
     return null;
   }
 
   return (
     <LayerList
+      ref={rightClickAreaRef}
       onPositionChange={(oldPos, newPos) =>
         arrangeElement({
           elementId: layers.find((layer) => layer.position === oldPos).id,
